# apps/backend/auth-service/src/controllers/auth.controller.ts
import configs from "@/src/config";
import {
  CreateUserRequest,
  GoogleCallbackRequest,
  LoginRequest,
  ResendVerifyCodeRequest,
  SignupRequest,
  VerifyUserRequest
} from "@/src/controllers/types/auth-request.type";
import AuthService from "@/src/services/auth.service";
import setCookie from "@/src/utils/cookie";
import { Response, Request as ExpressRequest } from "express";
import { Body, Controller, Get, Post, Queries, Request, Route } from "tsoa";

@Route("v1/auth")
export class AuthController extends Controller {
  @Get("/health")
  public async getHealth(): Promise<{ message: string }> {
    try {
      return { message: "OK" };
    } catch (error) {
      throw error;
    }
  }

  @Post("/create-user")
  public async createUser(
    @Body() body: CreateUserRequest
  ): Promise<{ message: string }> {
    try {
      const result = await AuthService.createUser(body);

      return { message: result };
    } catch (error) {
      console.error(`AuthController - createUser() method error: ${error}`);
      throw error;
    }
  }

  @Post("/signup")
  public async signup(
    @Body() body: SignupRequest
  ): Promise<{ message: string }> {
    try {
      const result = await AuthService.createUser(body);

      return { message: result };
    } catch (error) {
      console.error(`AuthController - signup() method error: ${error}`);
      throw error;
    }
  }

  @Post("/verify")
  public async verifyUser(@Body() body: VerifyUserRequest) {
    try {
      await AuthService.verifyUser(body);
      return { message: "You've verified successfully" };
    } catch (error) {
      console.error(`AuthController - verifyUser() method error: ${error}`);
      throw error;
    }
  }

  @Post("/resend-code")
  public async resendVerifyCode(@Body() body: ResendVerifyCodeRequest) {
    try {
      const result = await AuthService.resendVerifyCode(body);
      return { message: result };
    } catch (error) {
      console.error(
        `AuthController - resendVerifyCode() method error: ${error}`
      );
      throw error;
    }
  }

  @Post("/login")
  public async login(
    @Request() request: Express.Request,
    @Body() body: LoginRequest
  ) {
    try {
      const response = (request as any).res as Response;
      const result = await AuthService.login(body);

      setCookie(response, "id_token", result.idToken);
      setCookie(response, "access_token", result.accessToken);
      setCookie(response, "refresh_token", result.refreshToken, {
        maxAge: 30 * 24 * 3600 * 1000
      });
      setCookie(response, "username", result.username!, {
        maxAge: 30 * 24 * 3600 * 1000
      });
      setCookie(response, "user_id", result.userId!, {
        maxAge: 30 * 24 * 3600 * 1000
      });

      return { message: "Login successfully" };
    } catch (error) {
      console.error(`AuthController - login() method error: ${error}`);
      throw error;
    }
  }

  @Get("/google")
  public loginWithGoogle() {
    const cognitoOAuthURL = AuthService.loginWithGoogle();

    return {
      message: "Login with Google successfully",
      data: cognitoOAuthURL
    };
  }

  @Get("/oauth/callback")
  public async oauthCallback(
    @Request() request: Express.Request,
    @Queries() query: GoogleCallbackRequest
  ) {
    try {
      const response = (request as any).res as Response;
      const tokens = await AuthService.getOAuthToken(query);

      setCookie(response, "id_token", tokens.idToken);
      setCookie(response, "access_token", tokens.accessToken);
      setCookie(response, "refresh_token", tokens.refreshToken, {
        maxAge: 30 * 24 * 3600 * 1000
      });
      setCookie(response, "username", tokens.username!, {
        maxAge: 30 * 24 * 3600 * 1000
      });
      setCookie(response, "user_id", tokens.userId!, {
        maxAge: 30 * 24 * 3600 * 1000
      });

      response.redirect(configs.clientUrl);
    } catch (error) {
      console.error(`AuthController - oauthCallback() method error: ${error}`);
      throw error;
    }
  }

  @Post("/refresh-token")
  public async refreshToken(
    @Request() request: ExpressRequest,
    @Body() body: { refreshToken?: string; username?: string }
  ) {
    try {
      const response = (request as any).res as Response;

      const refreshToken = request.cookies["refresh_token"];
      const username = request.cookies["username"];

      const result = await AuthService.refreshToken({
        refreshToken: body.refreshToken || refreshToken,
        username: body.username || username
      });

      setCookie(response, "id_token", result.idToken);
      setCookie(response, "access_token", result.accessToken);

      return { message: "Token refreshed successfully" };
    } catch (error) {
      console.error(`AuthController - refreshToken() method error: ${error}`);
      throw error;
    }
  }
}